Bad_drivers/requirements.txt
Django
gunicorn
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
#django-versatileimagefield
drf-flex-fields
#pandas # its work
#Pillow==7.0